+++ src/hooks/sync-sites/sync-sites-context.tsx
@@ -6,7 +6,7 @@ import { useListenDeepLinkConnection } from './use-listen-deep-link-connection';
 import { UseSiteSyncManagement, useSiteSyncManagement } from './use-site-sync-management';
 import { PullStates, UseSyncPull, useSyncPull } from './use-sync-pull';
 import { PushStates, UseSyncPush, useSyncPush } from './use-sync-push';
-import type { SyncSite } from '../use-fetch-wpcom-sites/types/sync-site';
+import type { SyncSite } from '../use-fetch-wpcom-sites/types';
 
 type GetLastSyncTimeText = ( timestamp: string | null, type: 'pull' | 'push' ) => string;
 type UpdateSiteTimestamp = (
