+++ src/hooks/sync-sites/use-site-sync-management.ts
@@ -3,7 +3,7 @@ import { getIpcApi } from '../../lib/get-ipc-api';
 import { useAuth } from '../use-auth';
 import { FetchSites, useFetchWpComSites } from '../use-fetch-wpcom-sites';
 import { useSiteDetails } from '../use-site-details';
-import type { SyncSite } from '../use-fetch-wpcom-sites/types/sync-site';
+import type { SyncSite } from '../use-fetch-wpcom-sites/types';
 
 type ConnectedSites = SyncSite[];
 type LoadConnectedSites = () => Promise< void >;
