+++ src/hooks/sync-sites/use-sync-pull.ts
@@ -14,7 +14,7 @@ import {
 	UpdateState,
 	usePullPushStates,
 } from './use-pull-push-states';
-import type { SyncSite } from '../use-fetch-wpcom-sites/types/sync-site';
+import type { SyncSite } from '../use-fetch-wpcom-sites/types';
 
 export type SyncBackupState = {
 	remoteSiteId: number;
