+++ src/hooks/tests/reconcile-connected-sites.test.ts
@@ -1,5 +1,5 @@
 import { reconcileConnectedSites } from '../use-fetch-wpcom-sites/reconcile-connected-sites';
-import type { SyncSite } from '../use-fetch-wpcom-sites/types/sync-site';
+import type { SyncSite } from '../use-fetch-wpcom-sites/types';
 
 describe( 'reconcileConnectedSites', () => {
 	test( 'should update name, url, syncSupport properties', () => {
