--- src/hooks/use-fetch-wpcom-sites/types/sync-support.ts
@@ -1,7 +1,0 @@
-export type SyncSupport =
-	| 'unsupported'
-	| 'syncable'
-	| 'needs-transfer'
-	| 'already-connected'
-	| 'jetpack-site'
-	| 'deleted';
